requirements.txt
mcp>=0.1.0
httpx>=0.24.0
fastapi>=0.68.0
uvicorn>=0.15.0
pandas>=1.3.0
numpy>=1.21.0
python-dotenv>=0.19.0
requests>=2.26.0
aiohttp>=3.8.0
tenacity>=8.0.0
pydantic>=1.8.0
sqlite3